+++ atlas-data-storage/src/main/java/uk/ac/ebi/gxa/data/NetCDFMissingVal.java
@@ -20,7 +20,7 @@
  * http://gxa.github.com/gxa
  */
 
-package uk.ac.ebi.gxa.netcdf;
+package uk.ac.ebi.gxa.data;
 
 import ucar.ma2.DataType;
 import ucar.nc2.Attribute;
