+++ atlas-data-storage/src/main/java/uk/ac/ebi/gxa/data/NetCDFPredicates.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.gxa.netcdf;
+package uk.ac.ebi.gxa.data;
 
 import com.google.common.base.Function;
 import com.google.common.base.Predicate;
