+++ atlas-data-storage/src/test/java/uk/ac/ebi/gxa/netcdf/TestNetCDFProxy.java
@@ -20,7 +20,7 @@
  * http://gxa.github.com/gxa
  */
 
-package uk.ac.ebi.gxa.netcdf;
+package uk.ac.ebi.gxa.data;
 
 import junit.framework.TestCase;
