+++ atlas-loader/src/main/java/uk/ac/ebi/gxa/loader/service/AtlasMAGETABLoader.java
@@ -32,9 +32,9 @@ import uk.ac.ebi.gxa.loader.UnloadExperimentCommand;
 import uk.ac.ebi.gxa.loader.cache.AtlasLoadCache;
 import uk.ac.ebi.gxa.loader.dao.LoaderDAO;
 import uk.ac.ebi.gxa.loader.steps.*;
-import uk.ac.ebi.gxa.netcdf.NetCDFCreator;
-import uk.ac.ebi.gxa.netcdf.NetCDFCreatorException;
-import uk.ac.ebi.gxa.netcdf.AtlasDataDAO;
+import uk.ac.ebi.gxa.data.NetCDFCreator;
+import uk.ac.ebi.gxa.data.NetCDFCreatorException;
+import uk.ac.ebi.gxa.data.AtlasDataDAO;
 import uk.ac.ebi.gxa.utils.ZipUtil;
 import uk.ac.ebi.microarray.atlas.model.ArrayDesign;
 import uk.ac.ebi.microarray.atlas.model.Assay;
