+++ atlas-web/src/main/java/ae3/model/ExpressionMatrix.java
@@ -23,8 +23,8 @@
 package ae3.model;
 
 import uk.ac.ebi.gxa.exceptions.LogUtil;
-import uk.ac.ebi.gxa.netcdf.ExperimentWithData;
-import uk.ac.ebi.gxa.netcdf.AtlasDataException;
+import uk.ac.ebi.gxa.data.ExperimentWithData;
+import uk.ac.ebi.gxa.data.AtlasDataException;
 import uk.ac.ebi.microarray.atlas.model.ArrayDesign;
 
 /**
