+++ atlas-web/src/main/java/uk/ac/ebi/gxa/web/ui/plot/ExperimentPlot.java
@@ -25,11 +25,11 @@ package uk.ac.ebi.gxa.web.ui.plot;
 import com.google.common.base.Function;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
-import uk.ac.ebi.gxa.netcdf.ExpressionStatistics;
-import uk.ac.ebi.gxa.netcdf.FloatMatrixProxy;
-import uk.ac.ebi.gxa.netcdf.NetCDFDescriptor;
-import uk.ac.ebi.gxa.netcdf.NetCDFProxy;
-import uk.ac.ebi.gxa.netcdf.AtlasDataException;
+import uk.ac.ebi.gxa.data.ExpressionStatistics;
+import uk.ac.ebi.gxa.data.FloatMatrixProxy;
+import uk.ac.ebi.gxa.data.NetCDFDescriptor;
+import uk.ac.ebi.gxa.data.NetCDFProxy;
+import uk.ac.ebi.gxa.data.AtlasDataException;
 import uk.ac.ebi.gxa.utils.DoubleIndexIterator;
 import uk.ac.ebi.gxa.utils.FactorValueComparator;
 import uk.ac.ebi.microarray.atlas.model.UpDownExpression;
