+++ atlas-web/src/test/java/ae3/dao/ExperimentalDataTest.java
@@ -25,8 +25,8 @@ package ae3.dao;
 import ae3.model.ExperimentalData;
 import org.junit.After;
 import org.junit.Test;
-import uk.ac.ebi.gxa.netcdf.AtlasDataDAO;
-import uk.ac.ebi.gxa.netcdf.AtlasDataException;
+import uk.ac.ebi.gxa.data.AtlasDataDAO;
+import uk.ac.ebi.gxa.data.AtlasDataException;
 import uk.ac.ebi.gxa.web.filter.ResourceWatchdogFilter;
 import uk.ac.ebi.microarray.atlas.model.ArrayDesign;
 import uk.ac.ebi.microarray.atlas.model.Assay;
